% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{as_docx_table}
\alias{as_docx_table}
\title{Add the data table to a Word document}
\usage{
as_docx_table(x, doc, caption = NULL, lang = "en")
}
\arguments{
\item{x}{A ggplot object or the result of \code{\link[=alt_describe]{alt_describe()}}.}

\item{doc}{An \code{rdocx} object from \code{officer::read_docx()}.}

\item{caption}{Optional caption written before the table.}

\item{lang}{Language of the generated texts, if \code{x} is a plot.}
}
\value{
The \code{rdocx} object with the table added.
}
\description{
Requires the officer package.
}
